requirements.txt
ib_async>=0.9.0
pandas>=2.2.0
numpy>=1.26.4
jinja2>=3.1.3
matplotlib>=3.8.3
seaborn>=0.13.2
python-dotenv>=1.1.0
loguru>=0.7.2
backtrader>=1.9.76
pytz>=2024.1
currencyconverter>=0.5.0

# Web application dependencies
flask>=3.0.2
flask-cors>=4.0.0
gunicorn>=21.2.0
waitress>=2.1.2  # Windows-compatible WSGI server alternative to gunicorn
werkzeug>=3.0.1 
snaptrade-python-sdk>=1.0.0 
requests>=2.0.0
